lab2/requirements.txt
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
confluent-kafka==2.5.3
